refactor: retire legacy repository module
