+++ backend/app/repositories/account_data_repository.py
@@ -6,7 +6,6 @@ from ..db_models import User
 def delete_account(db: Session, user_id: str) -> None:
     user = db.get(User, user_id)
     if not user:
-        raise ValueError("?ъ슜???뺣낫瑜?李얠? 紐삵뻽?댁슂.")
+        raise ValueError("사용자 정보를 찾지 못했어요.")
     db.delete(user)
     db.commit()
-
