+++ src/hooks/useAppCoordinatorActionUtils.ts
@@ -2,7 +2,7 @@ export function formatCoordinatorErrorMessage(error: unknown) {
   if (error instanceof Error) {
     return error.message;
   }
-  return '?遺욧퍕??筌ｌ꼶???? 筌륁궢六??곸뒄. ?醫롫뻻 ??쇰퓠 ??쇰뻻 ??뺣즲??雅뚯눘苑??';
+  return '요청을 처리하지 못했어요. 조금 뒤에 다시 시도해 주세요.';
 }
 
 export function reportCoordinatorBackgroundError(error: unknown) {
